requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pandas
openpyxl
